requirements.txt
flask
typing
pydantic
langchain
python-dotenv